Added test for bfs_territory.
